Development
Methods to obtain the adjacent nodes of a node in terms of set of edges
pointing to an adjacent node, or a set of nodes reached by an edge.
Remove nodes in SimpleTreeImpl and SimpleGraphImpl. Bug in subtree
algorithm: a prune is needed to remove unnecessary nodes after the
traversal.